refactor: 그룹위치 변경 및 OpenWeatherMainViewController내부에 있던 CLLocationDelegate 역할 분리
