+++ WeatherForecast/WeatherForecast/Controller/OpenWeatherMainViewController.swift
@@ -9,20 +9,22 @@ import CoreLocation
 
 final class OpenWeatherMainViewController: UIViewController {
     private let locationManager = LocationManager()
- 
+    private let locationManagerDelegate = LocationManagerDelegate()
     private let tableView = UITableView()
     private let tableViewDataSource = WeatherTableviewDataSource()
     private let headerDelegate = WeatherTableViewDelegate()
-    
+
     //MARK: - View's Life Cycle
     override func viewDidLoad() {
         super.viewDidLoad()
         drawTableView()
         setRefreshControl()
-
+        
+        
         self.tableView.dataSource = self.tableViewDataSource
         self.tableView.delegate = headerDelegate
-        locationManager.delegate = self
+        
+        locationManager.delegate = locationManagerDelegate
         locationManager.askUserLocation()
         
         //MARK: Notified after OpenWeatherAPI response delivered successfully
@@ -50,56 +52,6 @@ final class OpenWeatherMainViewController: UIViewController {
     private let refreshControl = UIRefreshControl()
 }
 
-extension OpenWeatherMainViewController: CLLocationManagerDelegate {
-    func locationManager(_ manager: CLLocationManager,
-                         didUpdateLocations locations: [CLLocation]) {
-        guard let latitude = locations.last?.coordinate.latitude,
-              let longitude = locations.last?.coordinate.longitude else {
-            return
-        }
-        
-        let location: Location = (latitude, longitude)
-        let sessionDelegate = OpenWeatherSessionDelegate()
-        let networkManager = WeatherNetworkManager()
-        
-        networkManager.fetchOpenWeatherData(latitudeAndLongitude: location,
-                                            requestPurpose: .currentWeather,
-                                            sessionDelegate.session)
-        
-        networkManager.fetchOpenWeatherData(latitudeAndLongitude: location,
-                                            requestPurpose: .forecast,
-                                            sessionDelegate.session)
-    }
-    
-    func locationManager(_ manager: CLLocationManager,
-                         didFailWithError error: Error) {
-        if let error = error as? CLError {
-            switch error.code {
-            case .locationUnknown:
-                break
-            default:
-                print(error.localizedDescription)
-            }
-        }
-    }
-    
-    func locationManager(_ manager: CLLocationManager,
-                         didChangeAuthorization status: CLAuthorizationStatus) {
-        switch status {
-        case .restricted, .denied:
-            showAlert(title: "❌",
-                      message: "날씨 정보를 사용 할 수 없습니다.")
-            break
-        case .authorizedWhenInUse, .authorizedAlways, .notDetermined:
-            manager.requestLocation()
-            break
-        @unknown default:
-            showAlert(title: "🌟",
-                      message: "애플이 새로운 정보를 추가했군요! 확인 해 봅시다😄")
-        }
-    }
-}
-
 extension OpenWeatherMainViewController {
     private func setRefreshControl() {
         if #available(iOS 10.0, *) {
@@ -118,7 +70,6 @@ extension OpenWeatherMainViewController {
     @objc func stopRefesh() {
         DispatchQueue.main.async {
             self.tableView.refreshControl?.endRefreshing()
-
         }
     }
     
@@ -129,13 +80,11 @@ extension OpenWeatherMainViewController {
                                 forCellReuseIdentifier: "weatherCell")
         self.tableView.register(OpenWeatherHeaderView.self,
                                 forHeaderFooterViewReuseIdentifier: "weatherHeaderView")
-        let iconSize = 40
+        let iconSize = 50
         self.tableView.rowHeight = CGFloat(iconSize)
         
         let headerViewSize: CGFloat = 140
         self.tableView.sectionHeaderHeight = headerViewSize
-        self.tableView.backgroundColor = .clear
-        self.tableView.backgroundView?.addBackground(imageName: "cat")
     }
     
     private func showAlert(title: String, message: String) {
