+++ WeatherForecast/WeatherForecast/Models/TemperatureConverter.swift
@@ -21,5 +21,3 @@ struct TemperatureConverter {
         }
     }
 }
-
-
